% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychRstats.R
\name{download_exercises}
\alias{download_exercises}
\title{Download exercises for psychRstats}
\usage{
download_exercises(dir, force = FALSE)
}
\arguments{
\item{dir}{The directory defaults to your current working 
directory, but can be set otherwise.}

\item{force}{Whether to overwrite other folders if they exist
-- defaults to FALSE, but can overwrite by being set to TRUE.}
}
\description{
Wraps around the \code{\link[usethis]{use_zip}} function
to download the exercises. First checks whether the directory 
already exists. If you'd like to overwrite an existing directory,
you must use \code{force = TRUE}.
}
\examples{
\dontrun{
download_exercises()
download_exercises("~/Downloads/")
download_exercises(dir = "~/Downloads/", force = TRUE)
}
}
