% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychRstats.R
\name{lab}
\alias{lab}
\title{Load a tutorial from the psychRstats package using packages
\code{learnr} and, where appropriate, \code{gradethis}.}
\usage{
lab(name)
}
\arguments{
\item{name}{The name of the tutorial, in quotation marks. 
A full list of tutorials can be found by running.}
}
\description{
Load a tutorial from the psychRstats package using packages
\code{learnr} and, where appropriate, \code{gradethis}.
}
\examples{
\dontrun{
lab("intro-to-R")
}

}
