% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychRstats.R
\name{list_tutorials}
\alias{list_tutorials}
\title{Print all available tutorials in the psychRstats package.}
\usage{
list_tutorials(n = NULL)
}
\arguments{
\item{n}{How many rows to print.}
}
\description{
Include no arguments to print them all. You may pass an argument
to the print function, e.g., \code{list_tutorials(n = 2)} to 
limit the number of responses.
}
\examples{
list_tutorials()

}
