% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychRstats.R
\docType{package}
\name{psychRstats}
\alias{psychRstats}
\title{psychRstats: College-Level Intro Stats Exercises in R}
\description{
Based on Bard College's Statistics for Psychology (PSY 203, Fall
2020, taught by Assistant Professor of Psychology Justin 
Dainer-Best), these vignettes and exercises are intended to 
be used alongside an undergraduate course in statistics, or 
an understanding of statistics based on such a course. 
It includes a series of vignettes and exercises which build on 
one another to develop skills while also exploring basic 
statistical concepts like the t-test, ANOVA, or null-hypothesis 
statistical testing.
}
\section{psychRstats functions}{

There is only one added function, \code{\link{lab}}, which
will run a tutorial. It is just a mask of 
\code{learnr::run_tutorial()} with the package defaulting to 
this one.
For more information, look at the main github site,
\url{https://github.com/jdbest/r-psych-stats/}
}

