% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychRstats.R
\name{run_exercise}
\alias{run_exercise}
\title{Open an exercise for editing from the psychRstats package}
\usage{
run_exercise(exercise, dir)
}
\arguments{
\item{exercise}{The name or number of the exercise, e.g., 
\code{run_exercise(3)} or \code{run_exercise("03-exercise")}}

\item{dir}{The folder/directory to look for the exercise. Defaults
to your current working directory.}
}
\description{
Using the utils function \code{\link[utils]{file.edit}}, 
will open the file for 
editing based on the current working directory. 
Please note that if the file does not already exist, it 
will not be created; an error will be returned. In this case, 
you should search your computer for the correct directory, 
and consider changing R's working directory.
}
\examples{
\dontrun{
run_exercise("03-exercise")
run_exercise(3)
}

}
